requirements.txt
streamlit
matplotlib
seaborn
pandas
urlextract
wordcloud
Image
nltk
plotly.express
SentimentIntensityAnalyzer
re
Counter